+++ src/modules/manufacturing/entities/gis-header.vw.ts
@@ -16,9 +16,9 @@ export class GisHeaderView extends BaseEntity {
   @Column({ name: 'REPORT_DATE' })
   reportDate!: Date;
 
-  @Field()
+  @Field({ nullable: true })
   @Column({ name: 'ORDER_NO' })
-  orderNo!: string;
+  orderNo?: string;
 
   @Field()
   @Column({ name: 'PART_NO' })
@@ -91,6 +91,10 @@ export class GisHeaderView extends BaseEntity {
   @Field({ nullable: true })
   @Column({ name: 'TOTAL_METER' })
   totalMeter?: number;
+
+  @Field({ nullable: true })
+  @Column({ name: 'LOT_BATCH_NO' })
+  lotBatchNo?: string;
 
   @Field()
   @Column({ name: 'CREATED_BY' })